+++ substratevm/src/com.oracle.svm.graal/src/com/oracle/svm/graal/meta/SubstrateType.java
@@ -25,7 +25,7 @@
 package com.oracle.svm.graal.meta;
 
 import java.lang.annotation.Annotation;
-import java.util.ArrayList;
+import java.lang.reflect.Modifier;
 import java.util.Arrays;
 import java.util.Iterator;
 import java.util.NoSuchElementException;
@@ -42,9 +42,9 @@ import com.oracle.svm.core.meta.SharedType;
 import com.oracle.svm.core.meta.SubstrateObjectConstant;
 import com.oracle.svm.core.snippets.KnownIntrinsics;
 import com.oracle.svm.core.util.VMError;
+import com.oracle.truffle.api.nodes.DenyReplace;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.NodeClass;
-import com.oracle.truffle.api.nodes.NodeCloneable;
 
 import jdk.internal.misc.Unsafe;
 import jdk.vm.ci.meta.Assumptions.AssumptionResult;
@@ -473,54 +473,6 @@ public class SubstrateType extends NodeClass implements SharedType {
      * Implementation of Truffle NodeClass interface
      */
 
-    @SuppressWarnings("deprecation")
-    @Override
-    public com.oracle.truffle.api.nodes.NodeFieldAccessor getNodeClassField() {
-        return SubstrateNodeFieldAccessor.fromSubstrateField(
-                        getNodeFields(field -> DynamicHub.toClass(field.getDeclaringClass().getHub()) == Node.class && field.getName().equals("nodeClass")).iterator().next());
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public com.oracle.truffle.api.nodes.NodeFieldAccessor[] getCloneableFields() {
-        return nodeFieldIterableToArray(getNodeFields(field -> !SubstrateNodeFieldAccessor.isChildField(field) && !SubstrateNodeFieldAccessor.isChildrenField(field) &&
-                        NodeCloneable.class.isAssignableFrom(DynamicHub.toClass(field.getType().getHub()))));
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public com.oracle.truffle.api.nodes.NodeFieldAccessor[] getFields() {
-        return nodeFieldIterableToArray(getNodeFields(null));
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public com.oracle.truffle.api.nodes.NodeFieldAccessor getParentField() {
-        return SubstrateNodeFieldAccessor.fromSubstrateField(
-                        getNodeFields(field -> DynamicHub.toClass(field.getDeclaringClass().getHub()) == Node.class && field.getName().equals("parent")).iterator().next());
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public com.oracle.truffle.api.nodes.NodeFieldAccessor[] getChildFields() {
-        return nodeFieldIterableToArray(getNodeFields(field -> SubstrateNodeFieldAccessor.isChildField(field)));
-    }
-
-    @SuppressWarnings("deprecation")
-    @Override
-    public com.oracle.truffle.api.nodes.NodeFieldAccessor[] getChildrenFields() {
-        return nodeFieldIterableToArray(getNodeFields(field -> SubstrateNodeFieldAccessor.isChildrenField(field)));
-    }
-
-    @SuppressWarnings("deprecation")
-    private static com.oracle.truffle.api.nodes.NodeFieldAccessor[] nodeFieldIterableToArray(Iterable<SubstrateField> fields) {
-        ArrayList<com.oracle.truffle.api.nodes.NodeFieldAccessor> fieldList = new ArrayList<>();
-        for (SubstrateField field : fields) {
-            fieldList.add(SubstrateNodeFieldAccessor.fromSubstrateField(field));
-        }
-        return fieldList.toArray(new com.oracle.truffle.api.nodes.NodeFieldAccessor[0]);
-    }
-
     @Override
     public Iterator<Node> makeIterator(Node node) {
         return new SubstrateNodeIterator(node, this);
@@ -531,12 +483,6 @@ public class SubstrateType extends NodeClass implements SharedType {
     public Class<? extends Node> getType() {
         assert Node.class.isAssignableFrom(DynamicHub.toClass(getHub()));
         return (Class<? extends Node>) DynamicHub.toClass(getHub());
-    }
-
-    @Override
-    @SuppressWarnings("deprecation")
-    protected Iterable<SubstrateField> getNodeFields() {
-        return getNodeFields(null);
     }
 
     @Override
@@ -552,29 +498,25 @@ public class SubstrateType extends NodeClass implements SharedType {
         }
     }
 
-    private Iterable<SubstrateField> getNodeFields(Predicate<SubstrateField> filter) {
-        return () -> new SubstrateNodeFieldIterator(this, filter);
-    }
-
     @Override
     public void putFieldObject(Object field, Node receiver, Object value) {
         assert !getFieldType(field).isPrimitive();
         assert value == null || getFieldType(field).isInstance(value);
-        long offset = SubstrateNodeFieldAccessor.makeOffset((SubstrateField) field);
+        long offset = makeOffset((SubstrateField) field);
         Unsafe.getUnsafe().putObject(receiver, offset, value);
     }
 
     @Override
     public Object getFieldObject(Object field, Node receiver) {
         assert !getFieldType(field).isPrimitive();
-        long offset = SubstrateNodeFieldAccessor.makeOffset((SubstrateField) field);
+        long offset = makeOffset((SubstrateField) field);
         return Unsafe.getUnsafe().getObject(receiver, offset);
     }
 
     @Override
     public Object getFieldValue(Object field, Node node) {
         Class<?> fieldType = getFieldType(field);
-        long offset = SubstrateNodeFieldAccessor.makeOffset((SubstrateField) field);
+        long offset = makeOffset((SubstrateField) field);
         if (fieldType == boolean.class) {
             return Unsafe.getUnsafe().getBoolean(node, offset);
         } else if (fieldType == byte.class) {
@@ -598,12 +540,12 @@ public class SubstrateType extends NodeClass implements SharedType {
 
     @Override
     public boolean isChildField(Object field) {
-        return SubstrateNodeFieldAccessor.isChildField((SubstrateField) field);
+        return isChildField((SubstrateField) field);
     }
 
     @Override
     public boolean isChildrenField(Object field) {
-        return SubstrateNodeFieldAccessor.isChildrenField((SubstrateField) field);
+        return isChildrenField((SubstrateField) field);
     }
 
     @Override
@@ -612,24 +554,19 @@ public class SubstrateType extends NodeClass implements SharedType {
     }
 
     @Override
+    protected boolean isReplaceAllowed() {
+        boolean replaceDenied = Modifier.isFinal(getModifiers()) && getAnnotation(DenyReplace.class) != null;
+        return !replaceDenied;
+    }
+
+    @Override
     public Class<?> getFieldType(Object field) {
-        return SubstrateNodeFieldAccessor.makeType((SubstrateField) field);
+        return makeType((SubstrateField) field);
     }
 
     @Override
     public String getFieldName(Object field) {
         return ((SubstrateField) field).getName();
-    }
-}
-
-@SuppressWarnings("deprecation")
-class SubstrateNodeFieldAccessor extends com.oracle.truffle.api.nodes.NodeFieldAccessor.AbstractUnsafeNodeFieldAccessor {
-
-    private final long offset;
-
-    protected SubstrateNodeFieldAccessor(SubstrateField field, NodeFieldKind nodeFieldKind) {
-        super(nodeFieldKind, makeDeclaringClass(field), field.getName(), makeType(field));
-        this.offset = makeOffset(field);
     }
 
     protected static boolean isChildrenField(SubstrateField field) {
@@ -658,26 +595,6 @@ class SubstrateNodeFieldAccessor extends com.oracle.truffle.api.nodes.NodeFieldA
         return field.getLocation();
     }
 
-    @Override
-    public long getOffset() {
-        return offset;
-    }
-
-    static SubstrateNodeFieldAccessor fromSubstrateField(SubstrateField field) {
-        com.oracle.truffle.api.nodes.NodeFieldAccessor.NodeFieldKind nodeFieldKind;
-        if (DynamicHub.toClass(field.getDeclaringClass().getHub()) == Node.class && field.getName().equals("parent")) {
-            nodeFieldKind = com.oracle.truffle.api.nodes.NodeFieldAccessor.NodeFieldKind.PARENT;
-        } else if (DynamicHub.toClass(field.getDeclaringClass().getHub()) == Node.class && field.getName().equals("nodeClass")) {
-            nodeFieldKind = com.oracle.truffle.api.nodes.NodeFieldAccessor.NodeFieldKind.NODE_CLASS;
-        } else if (SubstrateNodeFieldAccessor.isChildField(field)) {
-            nodeFieldKind = com.oracle.truffle.api.nodes.NodeFieldAccessor.NodeFieldKind.CHILD;
-        } else if (SubstrateNodeFieldAccessor.isChildrenField(field)) {
-            nodeFieldKind = com.oracle.truffle.api.nodes.NodeFieldAccessor.NodeFieldKind.CHILDREN;
-        } else {
-            nodeFieldKind = com.oracle.truffle.api.nodes.NodeFieldAccessor.NodeFieldKind.DATA;
-        }
-        return new SubstrateNodeFieldAccessor(field, nodeFieldKind);
-    }
 }
 
 class SubstrateNodeFieldIterator implements Iterator<SubstrateField> {
@@ -783,13 +700,13 @@ class SubstrateNodeIterator implements Iterator<Node> {
     }
 
     private boolean computeNextFromField(SubstrateField field) {
-        if (SubstrateNodeFieldAccessor.isChildField(field)) {
+        if (SubstrateType.isChildField(field)) {
             long offset = field.getLocation();
             next = (Node) Unsafe.getUnsafe().getObject(node, offset);
             if (next != null) {
                 return true;
             }
-        } else if (SubstrateNodeFieldAccessor.isChildrenField(field)) {
+        } else if (SubstrateType.isChildrenField(field)) {
             long offset = field.getLocation();
             children = (Object[]) Unsafe.getUnsafe().getObject(node, offset);
             nextChildInChildren = 0;